requirements.txt
pandas
Flask
Flask-Cors
Flask-RESTful
mysql-connector-python
